+++ Mud.Importer.Rot/RotImporter.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using Mud.Domain;
+﻿using Mud.Domain;
 using Mud.Logger;
 using Mud.Server.Blueprints.Area;
 using Mud.Server.Blueprints.Character;
@@ -196,7 +195,7 @@ public class RotImporter
 
     private IRoomFlags ConvertRoomFlags(long input)
     {
-        var flags = ServiceProvider.GetRequiredService<IRoomFlags>();
+        var flags = new RoomFlags(ServiceProvider);
         if (IsSet(input, ROOM_DARK)) flags.Set("Dark");
         if (IsSet(input, ROOM_NO_MOB)) flags.Set("NoMob");
         if (IsSet(input, ROOM_INDOORS)) flags.Set("Indoors");
@@ -961,7 +960,7 @@ public class RotImporter
 
     private IItemFlags ConvertExtraFlags(ObjectData objectData)
     {
-        var itemFlags = ServiceProvider.GetRequiredService<IItemFlags>();
+        var itemFlags = new ItemFlags(ServiceProvider);
         if (IsSet(objectData.ExtraFlags, ITEM_GLOW)) itemFlags.Set("Glowing");
         if (IsSet(objectData.ExtraFlags, ITEM_HUM)) itemFlags.Set("Humming");
         if (IsSet(objectData.ExtraFlags, ITEM_DARK)) itemFlags.Set("Dark");
@@ -1030,7 +1029,7 @@ public class RotImporter
         }
 
         long weaponType2 = objectData.Values[4] == null ? 0L : System.Convert.ToInt64(objectData.Values[4]);
-        var weaponFlags = ServiceProvider.GetRequiredService<IWeaponFlags>();
+        var weaponFlags = new WeaponFlags(ServiceProvider);
         if (IsSet(weaponType2, WEAPON_FLAMING)) weaponFlags.Set("Flaming");
         if (IsSet(weaponType2, WEAPON_FROST)) weaponFlags.Set("Frost");
         if (IsSet(weaponType2, WEAPON_VAMPIRIC)) weaponFlags.Set("Vampiric");
@@ -1337,7 +1336,7 @@ public class RotImporter
 
     private IIRVFlags ConvertIRV(long value)
     {
-        var flags = ServiceProvider.GetRequiredService<IIRVFlags>();
+        var flags = new IRVFlags(ServiceProvider);
         if (IsSet(value, IMM_SUMMON)) flags.Set("Summon");
         if (IsSet(value, IMM_CHARM)) flags.Set("Charm");
         if (IsSet(value, IMM_MAGIC)) flags.Set("Magic");
@@ -1367,7 +1366,7 @@ public class RotImporter
 
     private ICharacterFlags ConvertCharacterFlags(long affectedBy)
     {
-        var flags = ServiceProvider.GetRequiredService<ICharacterFlags>();
+        var flags = new CharacterFlags(ServiceProvider);
         if (IsSet(affectedBy, AFF_BLIND)) flags.Set("Blind");
         if (IsSet(affectedBy, AFF_INVISIBLE)) flags.Set("Invisible");
         if (IsSet(affectedBy, AFF_DETECT_EVIL)) flags.Set("DetectEvil");
@@ -1403,7 +1402,7 @@ public class RotImporter
     }
     private IActFlags ConvertActFlags(long act, long act2)
     {
-        var flags = ServiceProvider.GetRequiredService<IActFlags>();
+        var flags = new ActFlags(ServiceProvider);
         //ACT_IS_NPC not used
         if (IsSet(act, ACT_SENTINEL)) flags.Set("Sentinel");
         if (IsSet(act, ACT_SCAVENGER)) flags.Set("Scavenger");
@@ -1441,7 +1440,7 @@ public class RotImporter
 
     private (IOffensiveFlags, IAssistFlags) ConvertOffensiveFlags(long input)
     {
-        var off = ServiceProvider.GetRequiredService<IOffensiveFlags>();
+        var off = new OffensiveFlags(ServiceProvider);
         if (IsSet(input, OFF_AREA_ATTACK)) off.Set("AreaAttack");
         if (IsSet(input, OFF_BACKSTAB)) off.Set("Backstab");
         if (IsSet(input, OFF_BASH)) off.Set("Bash");
@@ -1460,7 +1459,7 @@ public class RotImporter
         // OFF_FEED
         // OFF_CLAN_GUARD
 
-        var assist = ServiceProvider.GetRequiredService<IAssistFlags>();
+        var assist = new AssistFlags(ServiceProvider);
         if (IsSet(input, ASSIST_ALL)) assist.Set("All");
         if (IsSet(input, ASSIST_ALIGN)) assist.Set("Align");
         if (IsSet(input, ASSIST_RACE)) assist.Set("Race");
@@ -1474,7 +1473,7 @@ public class RotImporter
 
     private IBodyForms ConvertBodyForms(long input)
     {
-        var forms = ServiceProvider.GetRequiredService<IBodyForms>();
+        var forms = new BodyForms(ServiceProvider);
         if (IsSet(input, FORM_EDIBLE)) forms.Set("Edible");
         if (IsSet(input, FORM_POISON)) forms.Set("Poison");
         if (IsSet(input, FORM_MAGICAL)) forms.Set("Magical");
@@ -1507,7 +1506,7 @@ public class RotImporter
 
     private IBodyParts ConvertBodyParts(long input)
     {
-        var parts = ServiceProvider.GetRequiredService<IBodyParts>();
+        var parts = new BodyParts(ServiceProvider);
         if (IsSet(input, PART_HEAD)) parts.Set("Head");
         if (IsSet(input, PART_ARMS)) parts.Set("Arms");
         if (IsSet(input, PART_LEGS)) parts.Set("Legs");
