+++ Mud.Server.Interfaces/Admin/IAdminManager.cs
@@ -4,6 +4,8 @@ namespace Mud.Server.Interfaces.Admin;
 
 public interface IAdminManager
 {
+    void AddAdmin(IAdmin admin);
+    void RemoveAdmin(IAdmin admin);
     IAdmin? GetAdmin(ICommandParameter parameter, bool perfectMatch);
     IEnumerable<IAdmin> Admins { get; }
 }
