+++ Mud.Server.Interfaces/IWiznet.cs
@@ -4,6 +4,6 @@ namespace Mud.Server.Interfaces;
 
 public interface IWiznet
 {
-    void Wiznet(string message, WiznetFlags flags, AdminLevels minLevel = AdminLevels.Angel);
+    void Log(string message, WiznetFlags flags, AdminLevels minLevel = AdminLevels.Angel);
 
 }
