+++ Mud.Server.POC/Spells/SpellGiantSize.cs
@@ -1,0 +1,32 @@
+﻿using Mud.Domain;
+using Mud.Server.Ability;
+using Mud.Server.Ability.Spell;
+using Mud.Server.Affects;
+using Mud.Server.Interfaces.Ability;
+using Mud.Server.Interfaces.Affect;
+using Mud.Server.Interfaces.Aura;
+using Mud.Server.Random;
+
+namespace Mud.Server.POC.Spells;
+
+[Spell(SpellName, AbilityEffects.Buff)]
+public class SpellGiantSize : CharacterBuffSpellBase
+{
+    private const string SpellName = "Giant Size";
+
+    public SpellGiantSize(IRandomManager randomManager, IAuraManager auraManager)
+        : base(randomManager, auraManager)
+    {
+    }
+
+    protected override string SelfAlreadyAffectedMessage => "You are already affected.";
+    protected override string NotSelfAlreadyAffectedMessage => "{0:N} is already affected.";
+    protected override string VictimAffectMessage => "You are now affected by Giant Size.";
+    protected override string CasterAffectMessage => "{0:N} {0:b} now affected by Giant Size.";
+
+    protected override (int level, TimeSpan duration, IAffect[] affects) AuraInfo => (Caster.Level, TimeSpan.FromMinutes(5),
+        new IAffect[]
+        {
+            new CharacterSizeAffect{Value = Sizes.Giant},
+        });
+}
