+++ Mud.Server.Race/RaceManager.cs
@@ -1,43 +1,37 @@
-﻿using Mud.Common;
-using Mud.Server.Interfaces;
-using Mud.Server.Interfaces.Ability;
-using Mud.Server.Interfaces.Race;
+﻿using Mud.Server.Interfaces.Race;
 
 namespace Mud.Server.Race;
 
 public class RaceManager : IRaceManager
 {
-    private readonly List<IRace> _races;
+    private readonly Dictionary<string, IRace> _raceByNames;
 
-    public RaceManager(IAssemblyHelper assemblyHelper, IAbilityManager abilityManager)
+    public RaceManager(IEnumerable<IRace> races)
     {
-        // Get races using reflection
-        Type iRaceType = typeof (IRace);
-        Type iPlayableRaceType = typeof(IPlayableRace);
-        _races = assemblyHelper.AllReferencedAssemblies.SelectMany(a => a.GetTypes()
-            .Where(t => t.IsClass && !t.IsAbstract && iRaceType.IsAssignableFrom(t))
-            .Select(t => CreateInstance(t, iPlayableRaceType.IsAssignableFrom(t), abilityManager))) // don't use DependencyContainer
-            .Where(r => r != null)!
-            .ToList()!;
+        _raceByNames = new Dictionary<string, IRace>(StringComparer.InvariantCultureIgnoreCase);
+        foreach (var r in races)
+            _raceByNames.Add(r.Name, r);
     }
-    
+
     #region IRaceManager
 
     public IEnumerable<IPlayableRace> PlayableRaces
         => Races.OfType<IPlayableRace>();
 
     public IEnumerable<IRace> Races
-        => _races;
+        => _raceByNames.Values;
 
     public IRace? this[string name]
-        => _races.FirstOrDefault(x => StringCompareHelpers.StringEquals(x.Name, name));
+    {
+        get
+        {
+            if (string.IsNullOrWhiteSpace(name))
+                return null;
+            if (_raceByNames.TryGetValue(name, out var r))
+                return r;
+            return null;
+        }
+    }
 
     #endregion
-
-    private static IRace? CreateInstance(Type type, bool isPlayableRace, IAbilityManager abilityManager)
-    {
-        if (isPlayableRace)
-            return Activator.CreateInstance(type, abilityManager) as IRace;
-        return Activator.CreateInstance(type) as IRace;
-    }
 }
