+++ Mud.Server.WPFTestApplication/App.xaml.cs
@@ -79,17 +79,18 @@ public partial class App : Application
         services.AddSingleton<IAbilityManager, Ability.AbilityManager>();
         services.AddSingleton<IGameActionManager, GameAction.GameActionManager>();
         services.AddSingleton<ITimeManager, Server.TimeManager>();
-        services.AddSingleton<IWorld, World.World>();
         services.AddSingleton<IQuestManager, Quest.QuestManager>();
         services.AddSingleton<IAuraManager, Aura.AuraManager>();
         services.AddSingleton<IItemManager, Item.ItemManager>();
         services.AddSingleton<ICharacterManager, Character.CharacterManager>();
         services.AddSingleton<IRoomManager, Room.RoomManager>();
         services.AddSingleton<IAreaManager, Area.AreaManager>();
+        services.AddSingleton<IAdminManager, Admin.AdminManager>();
+        services.AddSingleton<IWiznet, Server.Wiznet>();
+        services.AddSingleton<IResetManager, Server.ResetManager>();
         services.AddSingleton<IServer, Server.Server>();
-        services.AddSingleton<IWiznet, Server.Server>(); // Server also implements IWiznet
+        services.AddSingleton<IWorld, Server.Server>(); // Server also implements IWorld
         services.AddSingleton<IPlayerManager, Server.Server>(); // Server also implements IPlayerManager
-        services.AddSingleton<IAdminManager, Server.Server>(); // Server also implements IAdminManager
         services.AddSingleton<IServerAdminCommand, Server.Server>(); // Server also implements IServerAdminCommand
         services.AddSingleton<IServerPlayerCommand, Server.Server>(); // Server also implements IServerPlayerCommand
         services.AddSingleton<IClassManager, Class.ClassManager>();
@@ -140,6 +141,20 @@ public partial class App : Application
         {
             Log.Default.WriteLine(LogLevels.Info, "Registering type {0}.", registrable.FullName);
             services.AddTransient(registrable);
+        }
+        // register races
+        var iRace = typeof(IRace);
+        foreach (var race in assemblies.SelectMany(a => a.GetTypes().Where(t => t.IsClass && !t.IsAbstract && iRace.IsAssignableFrom(t))))
+        {
+            Log.Default.WriteLine(LogLevels.Info, "Registering race type {0}.", race.FullName);
+            services.AddSingleton(iRace, race);
+        }
+        // register classes
+        var iClass = typeof(IClass);
+        foreach (var cl in assemblies.SelectMany(a => a.GetTypes().Where(t => t.IsClass && !t.IsAbstract && iClass.IsAssignableFrom(t))))
+        {
+            Log.Default.WriteLine(LogLevels.Info, "Registering class type {0}.", cl.FullName);
+            services.AddSingleton(iClass, cl);
         }
     }
