+++ Mud.Server/Admin/Administration/Slay.cs
@@ -41,7 +41,7 @@ public class Slay : AdminGameAction
 
     public override void Execute(IActionInput actionInput)
     {
-        Wiznet.Wiznet($"{Actor.DisplayName} slayed {Whom.DebugName}.", WiznetFlags.Punish);
+        Wiznet.Log($"{Actor.DisplayName} slayed {Whom.DebugName}.", WiznetFlags.Punish);
 
         Whom.Act(ActOptions.ToAll, "{0:N} slay{0:v} {1} in cold blood!", Actor.Impersonating!, Whom);
         Whom.RawKilled(Actor.Impersonating, false);
