+++ Mud.Server/Admin/Incarnate.cs
@@ -74,7 +74,7 @@ public class Incarnate : AdminGameAction
     {
         if (Target == null)
         {
-            Wiznet.Wiznet($"{Actor.DisplayName} stops incarnating {Actor.Incarnating?.DebugName}.", Domain.WiznetFlags.Incarnate);
+            Wiznet.Log($"{Actor.DisplayName} stops incarnating {Actor.Incarnating?.DebugName}.", Domain.WiznetFlags.Incarnate);
 
             Actor.Send("%M%You stop incarnating %C%{0}%x%.", Actor.Incarnating?.DisplayName ?? "???");
             Actor.StopIncarnating();
@@ -86,7 +86,7 @@ public class Incarnate : AdminGameAction
         if (incarnated)
         {
             string msg = $"{Actor.DisplayName} starts incarnating {Target.DebugName}.";
-            Wiznet.Wiznet(msg, Domain.WiznetFlags.Incarnate);
+            Wiznet.Log(msg, Domain.WiznetFlags.Incarnate);
 
             Actor.Send("%M%You start incarnating %C%{0}%x%.", Target.DisplayName);
         }
