+++ Mud.Server/Area/AreaManager.cs
@@ -1,21 +1,16 @@
 ﻿using Mud.Logger;
 using Mud.Server.Blueprints.Area;
-using Mud.Server.Interfaces;
 using Mud.Server.Interfaces.Area;
 
 namespace Mud.Server.Area;
 
 public class AreaManager : IAreaManager
 {
-    private IWiznet Wiznet { get; }
-
     private readonly Dictionary<int, AreaBlueprint> _areaBlueprints;
     private readonly List<IArea> _areas;
 
-    public AreaManager(IWiznet wiznet)
+    public AreaManager()
     {
-        Wiznet = wiznet;
-
         _areaBlueprints = [];
         _areas = [];
     }
@@ -41,7 +36,7 @@ public class AreaManager : IAreaManager
 
     public IArea AddArea(Guid guid, AreaBlueprint blueprint)
     {
-        IArea area = new Area(Wiznet, guid, blueprint);
+        IArea area = new Area(guid, blueprint);
         _areas.Add(area);
         return area;
     }
